+++ wdl-lsp/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-* Added `goto_definition` tests ([#488](https://github.com/stjude-rust-labs/wdl/pull/488)).
+* Added `goto_definition`, `find_all_references` tests ([#489](https://github.com/stjude-rust-labs/wdl/pull/489)).
 
 #### Fixed